+++ src/main/kotlin/com/dsoftware/ghtoolbar/data/WorkflowRunListLoader.kt
@@ -31,7 +31,7 @@ class WorkflowRunListLoader(
     }
 
     override fun reset() {
-        LOG.info("Removing all from the list model")
+        LOG.debug("Removing all from the list model")
         listModel.removeAll()
         loaded = false
 
@@ -46,13 +46,13 @@ class WorkflowRunListLoader(
     override fun canLoadMore() = !loading && !loaded
 
     override fun doLoadMore(indicator: ProgressIndicator, update: Boolean): List<GitHubWorkflowRun> {
-        LOG.info("Do load more update: $update, indicator: $indicator")
+        LOG.debug("Do load more update: $update, indicator: $indicator")
 
-        LOG.info("Get workflow runs")
+        LOG.debug("Get workflow runs")
         val request = Workflows.getWorkflowRuns(repositoryCoordinates)
         val response = requestExecutor.execute(indicator, request)
         val result = response.workflow_runs
-        LOG.info("Got ${result.size} workflows")
+        LOG.debug("Got ${result.size} workflows")
         //This is quite slow - N+1 requests, but there are no simpler way to get it, at least now.
 //        result.parallelStream().forEach {
 //            LOG.info("Get workflow by url ${it.workflow_url}")